refactor `M2MBase`
